+++ nn_dataflow/core/loop_blocking_scheme.py
@@ -218,7 +218,7 @@ class LoopBlockingScheme(object):
         indexed by DataCategoryEnum.
         '''
         if not self.is_valid():
-            return None
+            return [[float('inf')] * de.NUM for _ in range(me.NUM)]
 
         if not self.finalized_stats:
             self._calc_stats()
@@ -237,9 +237,9 @@ class LoopBlockingScheme(object):
 
         return self.fetch[self.BL.GBUF]
 
-    def get_cost(self, cost):
+    def get_access_cost(self, cost):
         '''
-        Get the total cost of loop blocking.
+        Get the data access cost of loop blocking.
         '''
         if not self.is_valid():
             return float('inf')
@@ -247,18 +247,10 @@ class LoopBlockingScheme(object):
         if not self.finalized_stats:
             self._calc_stats()
 
-        c = 0
+        acc_cost = sum(c * sum(a) for c, a in zip(cost.mem_hier, self.access))
+        acc_cost += cost.mem_hier_at(me.GBUF) * sum(self.remote_gbuf_access)
 
-        c += self.ops * cost.mac_op
-
-        access_total = [sum(acc) for acc in self.access]
-        c += sum(mc * ma for mc, ma in zip(cost.mem_hier, access_total))
-
-        c += sum(self.remote_gbuf_access) * cost.mem_hier_at(me.GBUF)
-
-        c += self.time * cost.unit_static * self.num_nodes
-
-        return c
+        return acc_cost
 
     def gen_index(self):
         '''