+++ nn_dataflow/core/nn_dataflow.py
@@ -79,10 +79,21 @@ class NNDataflow(object):
         # extended to the end of the network.
         self.nndf_tops = {}
 
+        # Default compare key function.
+        self.cmp_key = lambda nndf: (nndf.total_cost, nndf.total_time)
+
     def schedule_search(self, options):
         '''
         Search the optimized dataflows.
         '''
+        # Set key function.
+        if options.opt_goal == 'ed':
+            self.cmp_key = lambda nndf: nndf.total_cost * nndf.total_time
+        elif options.opt_goal == 'd':
+            self.cmp_key = lambda nndf: (nndf.total_time, nndf.total_cost)
+        else:
+            assert options.opt_goal == 'e'
+
         # Group the segments by the ending layers.
         segments = defaultdict(list)
         for seg in self.ilp.gen_segment(options):
@@ -116,7 +127,7 @@ class NNDataflow(object):
                 tops += self._segment_schedule_search(seg, options)
 
             # Always pick and keep top n.
-            tops = sorted(tops, key=NNDataflowScheme.cmp_key)[:options.ntops]
+            tops = sorted(tops, key=self.cmp_key)[:options.ntops]
 
             # Add to the top list.
             assert layer_name not in self.nndf_tops
@@ -223,7 +234,7 @@ class NNDataflow(object):
             nndf_tops += seg_nndf_tops
 
         # Always pick and keep top n.
-        return sorted(nndf_tops, key=NNDataflowScheme.cmp_key)[:options.ntops]
+        return sorted(nndf_tops, key=self.cmp_key)[:options.ntops]
 
     def _layer_schedule_search(self, layer_name, resource, constraint,
                                spatial_idx, temporal_idx, fwd_data_region,
@@ -243,8 +254,9 @@ class NNDataflow(object):
 
         layer_sched = self.layer_sched_dict[layer_name]
 
-        for ifmap_layout, prev_nndf in self._gen_layer_ifmap_layout(
-                layer_name, prev_nndf_tops):
+        for prev_nndf in prev_nndf_tops:
+
+            ifmap_layout = prev_nndf.fmap_layout(self.network.prevs(layer_name))
 
             if fwd_data_region is not None:
                 # Remap source data regions to the forwarding region.
@@ -283,32 +295,7 @@ class NNDataflow(object):
                 nndf_tops.append(nndf)
 
         # Always pick and keep top n at each layer.
-        return sorted(nndf_tops, key=NNDataflowScheme.cmp_key)[:options.ntops]
-
-    def _gen_layer_ifmap_layout(self, layer_name, prev_nndf_tops):
-        '''
-        Generate all choices of ifmap layout for the layer, based on the given
-        previous top NNDataflowScheme instances in `prev_nndf_tops`.
-
-        Return the ifmap layout, and the corresponding NNDataflowScheme.
-        '''
-        prevs = self.network.prevs(layer_name)
-        assert prevs
-
-        def _ofmap_layout(nndf, prev_layer_name):
-            return nndf[prev_layer_name].ofmap_layout if prev_layer_name \
-                    else nndf.input_layout
-
-        for nndf in prev_nndf_tops:
-            # Merge all previous layer ofmap layouts to get the ifmap layout.
-            ifmap_layout = DataLayout.concat(*[_ofmap_layout(nndf, p)
-                                               for p in prevs])
-
-            # We already checked the ofmap layout dimension in Scheduling, and
-            # the prev/next layer dimensions in Network, so ifmap_layout ==
-            # layer == prev_layers == ofmap_layout.
-
-            yield ifmap_layout, nndf
+        return sorted(nndf_tops, key=self.cmp_key)[:options.ntops]
 
     def _gen_input_layout(self, options):
         '''