+++ nn_dataflow/core/partition.py
@@ -61,6 +61,12 @@ def gen_partition(layer, batch_size, dim_nodes, options, guaranteed=False):
         elif batch_size % pdims[pe.BATP].size() != 0:
             continue
 
+        # Force each partitioning to fully utilize one dimension before
+        # startint the other, except for OFMP.
+        if any(1 < pd.h < dim_nodes.h and 1 < pd.w < dim_nodes.w
+               and pae != pe.OFMP for pae, pd in enumerate(pdims)):
+            continue
+
         if options.partition_hybrid:
             # Require partition is approximately dividable of total size.
             if not util.approx_dividable(layer.nofm, pdims[pe.OUTP].size()):
@@ -68,13 +74,6 @@ def gen_partition(layer, batch_size, dim_nodes, options, guaranteed=False):
             if not util.approx_dividable(layer.hofm, pdims[pe.OFMP].h) \
                     or not util.approx_dividable(layer.wofm, pdims[pe.OFMP].w):
                 continue
-
-            if options.partition_dimensional:
-                # Force each partitioning to be only along one dimension,
-                # except if there is only one partitioning.
-                if any(pd.h != 1 and pd.w != 1 for pd in pdims) \
-                        and sum(pd.size() > 1 for pd in pdims) > 1:
-                    continue
 
             if (not options.partition_ifmaps) and pdims[pe.INPP].size() > 1:
                 continue
@@ -273,12 +272,11 @@ def unit_nhops_to_proc_region(layer, batch_size, region, part,
         coord = part.coordinate(region, pidx)
         filrng, ifrng, ofrng = proc_data_range(layer, batch_size, part, pidx)
 
-        if not filrng[0].empty() and not filrng[1].empty():
-            fil_dict.setdefault(filrng, []).append(coord)
-        if ifrng.size() > 0:
+        if ifrng.size() > 0 and ofrng.size() > 0:
             ifm_dict.setdefault(ifrng, []).append(coord)
-        if ofrng.size() > 0:
             ofm_dict.setdefault(ofrng, []).append(coord)
+            if not filrng[0].empty() and not filrng[1].empty():
+                fil_dict.setdefault(filrng, []).append(coord)
 
     # All data should be processed by the same number of nodes, or no node.
     assert len(set(len(v) for v in fil_dict.values())) <= 1, \