+++ nn_dataflow/tests/dataflow_test/test_nn_dataflow.py
@@ -86,7 +86,7 @@ class TestNNDataflow(unittest.TestCase):
         self.cost = Cost(mac_op=1,
                          mem_hier=(200, 6, 2, 1),
                          noc_hop=0,
-                         unit_static=0)
+                         idl_unit=0)
 
         self.options = Option()
 
@@ -290,6 +290,60 @@ class TestNNDataflow(unittest.TestCase):
         self.assertIs(nnd.layer_sched_dict['pool1_a'],
                       nnd.layer_sched_dict['pool1_b'])
 
+    def test_opt_goal(self):
+        ''' Optimization goal. '''
+        network = self.alex_net
+
+        batch_size = 8
+
+        resource = self.resource._replace(
+            proc_region=NodeRegion(origin=PhyDim2(0, 0),
+                                   dim=PhyDim2(8, 8),
+                                   type=NodeRegion.PROC)
+        )
+
+        nnd = NNDataflow(network, batch_size, resource, self.cost,
+                         self.map_strategy)
+
+        options_e = Option(sw_gbuf_bypass=(True, True, True),
+                           sw_solve_loopblocking=True,
+                           partition_hybrid=True,
+                           partition_batch=True,
+                           opt_goal='e',
+                           ntops=16)
+        tops_e, _ = nnd.schedule_search(options_e)
+        self.assertTrue(tops_e)
+
+        options_d = Option(sw_gbuf_bypass=(True, True, True),
+                           sw_solve_loopblocking=True,
+                           partition_hybrid=True,
+                           partition_batch=True,
+                           opt_goal='d',
+                           ntops=16)
+        tops_d, _ = nnd.schedule_search(options_d)
+        self.assertTrue(tops_d)
+
+        options_ed = Option(sw_gbuf_bypass=(True, True, True),
+                            sw_solve_loopblocking=True,
+                            partition_hybrid=True,
+                            partition_batch=True,
+                            opt_goal='ed',
+                            ntops=16)
+        tops_ed, _ = nnd.schedule_search(options_ed)
+        self.assertTrue(tops_ed)
+
+        self.assertLess(tops_e[0].total_cost, tops_d[0].total_cost)
+        self.assertLess(tops_e[0].total_cost, tops_ed[0].total_cost)
+
+        self.assertLess(tops_d[0].total_time, tops_e[0].total_time)
+        self.assertLess(tops_d[0].total_time, tops_ed[0].total_time)
+
+        # Sum of the smallest ED may not be the smallest; allow for error.
+        self.assertLess(tops_ed[0].total_cost * tops_ed[0].total_time,
+                        tops_e[0].total_cost * tops_e[0].total_time * 1.05)
+        self.assertLess(tops_ed[0].total_cost * tops_ed[0].total_time,
+                        tops_d[0].total_cost * tops_d[0].total_time * 1.05)
+
     def test_no_valid_dataflow(self):
         ''' No valid dataflow is found. '''
 
@@ -447,7 +501,7 @@ class TestNNDataflow(unittest.TestCase):
         cost = Cost(mac_op=2e-12,
                     mem_hier=(460e-12, 15e-12, 4e-12, 1e-12),  # pJ/16-b
                     noc_hop=0,
-                    unit_static=30e-3 / 200e6)  # 30 mW GBUF + REGF
+                    idl_unit=30e-3 / 200e6)  # 30 mW GBUF + REGF
 
         nnd = NNDataflow(network, batch_size, resource, cost,
                          self.map_strategy)
@@ -543,7 +597,7 @@ class TestNNDataflow(unittest.TestCase):
         cost = Cost(mac_op=2e-12,
                     mem_hier=(240e-12, 28e-12, 4e-12, 1e-12),  # pJ/16-b
                     noc_hop=0,
-                    unit_static=320e-12)
+                    idl_unit=320e-12)
 
         nnd = NNDataflow(network, batch_size, resource, cost,
                          self.map_strategy)
@@ -575,7 +629,7 @@ class TestNNDataflow(unittest.TestCase):
         cost = Cost(mac_op=2e-12,
                     mem_hier=(80e-12, 14e-12, 4e-12, 0.6e-12),  # pJ/16-b
                     noc_hop=40e-12,
-                    unit_static=200e-12)
+                    idl_unit=200e-12)
 
         options = Option(sw_gbuf_bypass=(True, True, True),
                          sw_solve_loopblocking=True,
@@ -596,5 +650,7 @@ class TestNNDataflow(unittest.TestCase):
         self.assertLess(dfsch_t16.total_time,
                         1.2 * dfsch_l1.total_time * (16 * 16) / (14 * 14 * 16))
         # Energy reduced by > 30%.
-        self.assertLess(dfsch_t16.total_cost, dfsch_l1.total_cost * 0.7)
+        # self.assertLess(dfsch_t16.total_cost, dfsch_l1.total_cost * 0.7)
+        # With dimension restriction on partitioning, this is slightly violated.
+        self.assertLess(dfsch_t16.total_cost, dfsch_l1.total_cost * 0.72)