+++ nn_dataflow/tests/dataflow_test/test_scheduling.py
@@ -48,7 +48,7 @@ class TestScheduling(unittest.TestCase):
         self.batch_size = 4
 
         self.cost = Cost(mac_op=1, mem_hier=(200, 6, 2, 1),
-                         noc_hop=50, unit_static=50)
+                         noc_hop=50, idl_unit=50)
 
         self.none_cstr = SchedulingConstraint()
         self.cstr = SchedulingConstraint(topofm=1, topbat=self.batch_size)
@@ -143,8 +143,10 @@ class TestScheduling(unittest.TestCase):
             # Combination of loop blocking and partitioning.
             for r in res:
                 self.assertAlmostEqual(r.total_cost,
-                                       r.scheme['cost_loop']
-                                       + r.scheme['cost_part'])
+                                       r.scheme['cost_op']
+                                       + r.scheme['cost_access']
+                                       + r.scheme['cost_noc']
+                                       + r.scheme['cost_static'])
                 self.assertEqual(r.total_ops, layer.total_ops(self.batch_size))
                 self.assertSequenceEqual(r.scheme['total_nhops'],
                                          [nh * f for nh, f
@@ -276,7 +278,7 @@ class TestScheduling(unittest.TestCase):
         self.assertNotEqual(id(opts), id(self.options))
 
         part = PartitionScheme(order=(pe.BATP, pe.INPP, pe.OUTP, pe.OFMP),
-                               pdims=((2, 2), (2, 2), (1, 1), (1, 1)))
+                               pdims=((2, 4), (2, 1), (1, 1), (1, 1)))
 
         _ = schd.schedule_search_per_node(part, rsrc, cstr, opts)